alertas con sweet alert

+++ client/src/actions/index.js
@@ -1,5 +1,9 @@
 import axios from "axios";
+import Swal from "sweetalert2";
+import withReactContent from "sweetalert2-react-content";
 import { FETCH_USER, FETCH_SURVEYS } from "./types";
+
+const MySwal = withReactContent(Swal);
 
 export const fetchUser = () => async (dispatch) => {
   const res = await axios.get("/api/current_user");
@@ -8,16 +12,42 @@ export const fetchUser = () => async (dispatch) => {
 
 export const handleToken = (token) => async (dispatch) => {
   const res = await axios.post("/api/stripe", token);
-
-  dispatch({ type: FETCH_USER, payload: res.data });
+  MySwal.fire({
+    title: <p>Realizando pago</p>,
+    didOpen: () => {
+      MySwal.showLoading();
+    },
+    timer: 2000,
+  }).then(() => {
+    MySwal.fire({
+      title: <p>Pago realizado correctamente.!</p>,
+      icon: "success",
+    });
+    return dispatch({ type: FETCH_USER, payload: res.data });
+  });
 };
 
 export const submitSurvey = (values, history) => async (dispatch) => {
-  const res = await axios.post("/api/surveys", values);
+  const sendMail = await MySwal.fire({
+    title: "¿Enviar correo?",
+    icon: "question",
+    showCancelButton: true,
+  });
+  if (sendMail.isConfirmed) {
+    const res = await axios.post("/api/surveys", values);
+    MySwal.fire({
+      title: "Enviando encuesta y redireccionando a tus encuestas",
+      icon: "info",
+      didOpen: () => {
+        MySwal.showLoading();
+      },
+      timer: 2000,
+    }).then(() => {
+      history.push("/surveys");
 
-  history.push("/surveys");
-
-  dispatch({ type: FETCH_USER, payload: res.data });
+      return dispatch({ type: FETCH_USER, payload: res.data });
+    });
+  }
 };
 
 export const fetchSurveys = () => async (dispatch) => {
@@ -25,8 +55,24 @@ export const fetchSurveys = () => async (dispatch) => {
 
   dispatch({ type: FETCH_SURVEYS, payload: res.data });
 };
-export const deleteSurvey = (surveyId) => async (dispatch) =>{
-  const res = await axios.delete("/api/surveys/delete/"+surveyId);
-
-  dispatch({ type: FETCH_SURVEYS, payload: res.data});
-} 
+export const deleteSurvey = (surveyId) => async (dispatch) => {
+  const deleteUser = await MySwal.fire({
+    title: "Eliminar Encuesta",
+    inputLabel: "¿Seguro que quieres eliminar esta encuesta?",
+    icon: "question",
+    showCancelButton: true,
+  });
+  if (deleteUser.isConfirmed) {
+    const res = await axios.delete("/api/surveys/delete/" + surveyId);
+    MySwal.fire({
+      title: "Eliminando encuesta",
+      icon: "info",
+      didOpen: () => {
+        MySwal.showLoading();
+      },
+      timer: 2000,
+    }).then(() => {
+      return dispatch({ type: FETCH_SURVEYS, payload: res.data });
+    });
+  }
+};
